requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2

# Database
supabase==2.9.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# Twilio
twilio==9.3.7

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.1

# Logging
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0

python-multipart